Move migration and model into the adapter, remove the initializer
