--- app/models/action_cable_record.rb
@@ -1,5 +1,0 @@
-class ActionCableRecord < ApplicationRecord
-  self.abstract_class = true
-
-  connects_to database: { writing: :action_cable, reading: :action_cable }
-end
