--- config/initializers/action_cable.rb
@@ -1,5 +1,0 @@
-ActiveSupport.on_load :action_cable do
-  def self.pubsub_adapter
-    SqliteSubscriptionAdapter
-  end
-end
