--- db/action_cable_migrate/20230318133745_create_action_cable_messages.rb
@@ -1,8 +1,0 @@
-class CreateActionCableMessages < ActiveRecord::Migration[7.0]
-  def change
-    create_table :messages do |t|
-      t.string :channel
-      t.blob   :message
-    end
-  end
-end
